Update compute part with external-beta situation
## GET/images


```
Images can also be created from instances or imported from another source.
```
==> External beta does not support create image. Please remove this words.

```
Required Parameters:
start: A server-supplied token determining what resource to start the page on.
limit: The number of resources to return on a page.
resource_group.id: Filters the collection to resources within one of the resource groups identified in a comma-separated list of resource group identifiers.
```
==> these are not required parameters, and not supported in external beta. 

## GET/images/{id}

```
400: An image with the specified identifier could not be found.
```
==> The code is 404

## GET/instance/profiles

```
Required Parameters:
start: A server-supplied token determining what resource to start the page on.
limit: The number of resources to return on a page.
```
==> They are optional query, not required parameters

## GET/instances

```
Required Parameters:
start: A server-supplied token determining what resource to start the page on.
limit: The number of resources to return on a page.
resource_group.id: Filters the collection to resources within one of the resource groups identified in a comma-separated list of resource group identifiers.
network_interfaces.subnet.id: Filters the collection to instances on the subnet of the specified identifier.
network_interfaces.subnet.cm: Filters the collection to instances on the subnet of the specified CRN.
network_interfaces.subnet.name: Filters the collection to instances on the subnet of the specified name.
```
==> support `start` and `limit` query, but they are not required. Don't support others.

## POST/instances

```
{
  "image": {
    "name": "my-image"
  },
  "keys": [
    {
      "fingerprint": "yxavE4CIOL2NlsqcurRO3xGjkP6m/0mp8ugojH5yxlY"
    }
  ],
  "profile": {
    "name": "gc.balanced.4x16"
  },
  "vpc": {
    "name": "my-vpc"
  }
}
```
==>

```
{
  "name": "longrun-test-1529390097",
  "type": "virtual",
    "zone": {
    "name": "us-south-1"
  },
  "flavor": {
    "name": "b-2x4"
  },
  "image": {
    "id": "7eb4e35b-4257-56f8-d7da-326d85452591"
  },
  "vpc": {
    "id": "fabc1711-9330-4a25-b44f-901c0710b5bb"
  },
  "primary_network_interface": {
    "port_speed": 1000,
    "security_groups": [],
    "subnet": {
      "id": "3d9a8ba9-ac74-4093-985c-be9876b4f645"
    }
  },
  "keys": [
    {
      "id": "8cf60ce4-9115-4dbe-a493-0f82a9eb2a97"
    }
  ],
  "user_data": ""
}
```

UI team don't agree change `flavor` to `profile` for now. https://github.ibm.com/riaas/cland/issues/702

## PATCH/instances/{id}

This API is not supported in external beta, see api-spec issue:  https://github.ibm.com/riaas/api-spec/issues/391

## GET/instances/{instance_id}/actions

```
Required Parameters:
instance_id: The instance identifier.
start: A server-supplied token determining what resource to start the page on.
limit: The number of resources to return on a page.
```
 ==>
`start` and `limit` are optional not required.

## POST/instances/{instance_id}/actions

There is another return code 409, that means the the action cannot be perform on the server. For example, you want to perform action on `Failed` or `Deleting` server, or you want to perform `stop` action on a `Stopped` server.

## GET/instances/{instance_id}/network_interfaces

`resource_group.id` is not supported

## PATCH/instances/{instance_id}/network_interfaces/{id}

This API is not supported. See api-spec issue https://github.ibm.com/riaas/api-spec/issues/418

## DELETE/instances/{instance_id}/network_interfaces/{network_interface_id}/floating_ips/{id}

This API is not supported, see api-spec issue https://github.ibm.com/riaas/api-spec/issues/415

## PUT/instances/{instance_id}/network_interfaces/{network_interface_id}/floating_ips/{id}

This API is not supported, see api-spec issue https://github.ibm.com/riaas/api-spec/issues/415

## GET/keys

```
Required Parameters:
start: A server-supplied token determining what resource to start the page on.
limit: The number of resources to return on a page.
resource_group.id: Filters the collection to resources within one of the resource groups identified in a comma-separated list of resource group identifiers.
```
==>
These query parameters are not supported on external beta

## POST/keys

```
  "resource_group": {
    "id": "56969d60-43e9-465c-883c-b9f7363e78e8"
  },
```
==>
Remove this part, since we don't support resource_group